folder refactoring for user task
