+++ src/api/routes/router.ts
@@ -34,7 +34,7 @@ import { register as registerPatientHealthProfileRoutes } from './patient/health
 import { register as registerPatientRoutes } from "./patient/patient.routes";
 import { register as registerPersonRoutes } from './person.routes';
 import { register as registerTypesRoutes } from './types.routes';
-import { register as registerUserTaskRoutes } from './user.task.routes';
+import { register as registerUserTaskRoutes } from './user/user.task.routes';
 import { register as registerUserDeviceDetailsRoutes } from './user/user.device.details.routes';
 import { register as registerUserRoutes } from "./user/user.routes";
 import { register as registerCalorieBalanceRoute } from './wellness/daily.records/calorie.balance.routes';
